requirements.txt
alembic==1.4.2
autopep8==1.5.2
beautifulsoup4==4.9.1
blinker==1.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
flake8==3.8.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Scss==0.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jac==0.17.2
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
ordereddict==1.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyScss==1.3.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
rjsmin==1.1.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
